+++ fabric/src/main/java/io/github/mortuusars/scholar/fabric/BookItemColorsReloadListenerFabric.java
@@ -1,10 +1,9 @@
 package io.github.mortuusars.scholar.fabric;
 
-import io.github.mortuusars.scholar.client.chiseled_bookshelf.BookshelfItemColorsReloadListener;
 import net.fabricmc.fabric.api.resource.IdentifiableResourceReloadListener;
 import net.minecraft.resources.Identifier;
 
-public class BookshelfItemColorsReloadListenerFabric extends BookshelfItemColorsReloadListener implements IdentifiableResourceReloadListener {
+public class BookItemColorsReloadListenerFabric extends BookItemColorsReloadListener implements IdentifiableResourceReloadListener {
     @Override
     public Identifier getFabricId() {
         return ID;